feat: work on Section and restructure ClassDetail page
